requirements.txt
langchain
openai
chromadb
youtube-transcript-api
beautifulsoup4
requests
gspread
oauth2client
pyyaml
